Code review comment: split ServiceDefaults extensions class into multiple.
