+++ Enigmatry.Entry.Blueprint.ServiceDefaults/ProgramExtensions.cs
@@ -1,19 +1,7 @@
-﻿using System.Reflection;
-using Azure.Monitor.OpenTelemetry.AspNetCore;
-using Azure.Monitor.OpenTelemetry.Exporter;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Diagnostics.HealthChecks;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Diagnostics.HealthChecks;
+﻿using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using OpenTelemetry;
-using OpenTelemetry.Metrics;
-using OpenTelemetry.Resources;
-using OpenTelemetry.Trace;
 
-namespace Enigmatry.Entry.ServiceDefaults;
+namespace Enigmatry.Entry.Blueprint.ServiceDefaults;
 
 // Adds common .NET Aspire services: service discovery, resilience, health checks, and OpenTelemetry.
 // This project should be referenced by each service project in your solution.
@@ -39,127 +27,4 @@ public static class ProgramExtensions
 
         return builder;
     }
-
-    public static IHostApplicationBuilder ConfigureOpenTelemetry(this IHostApplicationBuilder builder)
-    {
-        builder.Logging.ConfigureOpenTelemetryLogging(builder.Configuration);
-
-        builder.Services.AddOpenTelemetry()
-            .WithMetrics(metrics =>
-            {
-                metrics.AddAspNetCoreInstrumentation()
-                    .AddHttpClientInstrumentation()
-                    .AddRuntimeInstrumentation();
-            })
-            .WithTracing(tracing =>
-            {
-                tracing.AddAspNetCoreInstrumentation()
-                    // Uncomment the following line to enable gRPC instrumentation (requires the OpenTelemetry.Instrumentation.GrpcNetClient package)
-                    //.AddGrpcClientInstrumentation()
-                    .AddHttpClientInstrumentation();
-                    // Uncomment the following line to enable sampling on collecting traces. 
-                    //.SetSampler(new TraceIdRatioBasedSampler(0.5));
-
-                if (builder.Environment.IsDevelopment())
-                {
-                    tracing.AddSqlClientInstrumentation(options =>
-                    {
-                        options.SetDbStatementForText = true;
-                    });
-                }
-            })
-            .AddOpenTelemetryExporters(builder.Configuration);
-
-        return builder;
-    }
-
-    public static IServiceCollection AddOpenTelemetryWorkerService(this IServiceCollection services, IConfiguration configuration)
-    {
-        var serviceName = Assembly.GetExecutingAssembly().FullName!;
-        services.AddOpenTelemetry()
-            .WithTracing(tracing =>
-            {
-                if (IsAzureMonitorConfigured(configuration))
-                {
-                    tracing
-                        .SetResourceBuilder(ResourceBuilder.CreateDefault()
-                        .AddService(serviceName))
-                        .AddSource(serviceName)
-                        .AddAzureMonitorTraceExporter();
-                }
-            })
-            .WithMetrics(metrics =>
-            {
-                if (IsAzureMonitorConfigured(configuration))
-                {
-                    metrics.AddAzureMonitorMetricExporter();
-                }
-            });
-
-        return services;
-    }
-
-    public static void ConfigureOpenTelemetryLogging(this ILoggingBuilder loggingBuilder, IConfiguration configuration)
-    {
-        // Clear all logging providers and add OpenTelemetry logging provider
-        loggingBuilder.ClearProviders();
-        loggingBuilder.AddOpenTelemetry(options =>
-        {
-            options.IncludeFormattedMessage = true;
-            options.IncludeScopes = true;
-            if (IsAzureMonitorConfigured(configuration))
-            {
-                options.AddAzureMonitorLogExporter();
-            }
-        });
-    }
-
-    private static void AddOpenTelemetryExporters(this OpenTelemetryBuilder builder, IConfiguration configuration)
-    {
-        if (IsOtlpCollectorConfigured(configuration))
-        {
-            builder.Services.AddOpenTelemetry().UseOtlpExporter();
-        }
-
-        // Uncomment the following lines to enable the Azure Monitor exporter (requires the Azure.Monitor.OpenTelemetry.AspNetCore package)
-        if (IsAzureMonitorConfigured(configuration))
-        {
-            builder.Services.AddOpenTelemetry()
-                .UseAzureMonitor();
-        }
-    }
-    public static IHostApplicationBuilder AddDefaultHealthChecks(this IHostApplicationBuilder builder)
-    {
-        builder.Services.AddHealthChecks()
-            // Add a default liveness check to ensure app is responsive
-            .AddCheck("self", () => HealthCheckResult.Healthy(), ["live"]);
-
-        return builder;
-    }
-
-    public static WebApplication MapDefaultEndpoints(this WebApplication app)
-    {
-        // Adding health checks endpoints to applications in non-development environments has security implications.
-        // See https://aka.ms/dotnet/aspire/healthchecks for details before enabling these endpoints in non-development environments.
-        if (app.Environment.IsDevelopment())
-        {
-            // All health checks must pass for app to be considered ready to accept traffic after starting
-            app.MapHealthChecks("/health");
-
-            // Only health checks tagged with the "live" tag must pass for app to be considered alive
-            app.MapHealthChecks("/alive", new HealthCheckOptions
-            {
-                Predicate = r => r.Tags.Contains("live")
-            });
-        }
-
-        return app;
-    }
-
-    private static bool IsAzureMonitorConfigured(IConfiguration configuration)
-        => !String.IsNullOrEmpty(configuration["APPLICATIONINSIGHTS_CONNECTION_STRING"]);
-
-    private static bool IsOtlpCollectorConfigured(IConfiguration configuration)
-        => !String.IsNullOrWhiteSpace(configuration["OTEL_EXPORTER_OTLP_ENDPOINT"]);
-
 }
